Add support for complex attributes
By complex attributes we refer to attributes that are not as simple as just
storing a value.  It may refer to a collection or any other object whose
functionality is beyond merely storing and retrieving the value from Redis.
Notable examples beyond collections are the counter and the lock types.